Moved compatiblity shims to a distinct module.
