+++ mutant/contrib/related/management/__init__.py
@@ -4,9 +4,9 @@ from django.db.models import Q
 from django.db.models.fields.related import RelatedField
 from django.utils.six import string_types
 
+from ....compat import clear_opts_related_cache
 from ....db.models import MutableModel
 from ....models import ModelDefinition
-from ....utils import clear_opts_related_cache
 
 
 def mutable_model_prepared(signal, sender, definition, existing_model_class,
