+++ tests/test_model_defs.py
@@ -10,6 +10,7 @@ from django.db.utils import IntegrityError
 from django.test.utils import CaptureQueriesContext
 from django.utils.translation import ugettext as _
 
+from mutant.compat import clear_opts_related_cache, get_related_model
 from mutant.contrib.related.models import ForeignKeyDefinition
 from mutant.contrib.text.models import CharFieldDefinition
 from mutant.db.models import MutableModel
@@ -17,9 +18,7 @@ from mutant.models.model import (
     BaseDefinition, ModelDefinition, MutableModelProxy,
     OrderingFieldDefinition, UniqueTogetherDefinition,
 )
-from mutant.utils import (
-    clear_opts_related_cache, get_related_model, remove_from_app_cache,
-)
+from mutant.utils import remove_from_app_cache
 
 from .models import (
     AbstractConcreteModelSubclass, AbstractModel, Mixin,
